pyproject.toml
[tool.poetry]
name = "mcgill-courses"
version = "0.1.0"
description = ""
authors = ["mcgill.courses <admin@mcgill.courses>"]
readme = "README.md"
packages = [{ include = "mcgill" }]

[tool.poetry.dependencies]
python = "^3.10"
llama-index = "^0.7.6"
python-dotenv = "^1.0.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
